backend/requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
sse-starlette>=2.0.0
